+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+[1.3.0] - 2024-11-22
+
+### Added
+- Added support for Personalise rules within `Mutually Exclusive Groups`.
+
 [1.2.0] - 2024-06-17
 
 ### Fixed
@@ -69,7 +74,7 @@ Send event properties as third param in trackEvent() call`
 
     @Override
     public void set(Map<String, Object> data) throws Exception {
-        String key = data.get("featureKey") + "_" + data.get("user");
+        String key = data.get("featureKey") + "_" + data.get("userId");
 
         // Create a map to store the data
         Map<String, Object> value = new HashMap<>();